[REFACTORING] Split IMAPServerPlainAuthenticateThenAnotherCommandTest
