+++ server/protocols/protocols-imap4/src/test/java/org/apache/james/imapserver/netty/IMAPServerPlainAuthenticateThenAnotherCommandTest.java
@@ -1,0 +1,66 @@
+/****************************************************************
+ * Licensed to the Apache Software Foundation (ASF) under one   *
+ * or more contributor license agreements.  See the NOTICE file *
+ * distributed with this work for additional information        *
+ * regarding copyright ownership.  The ASF licenses this file   *
+ * to you under the Apache License, Version 2.0 (the            *
+ * "License"); you may not use this file except in compliance   *
+ * with the License.  You may obtain a copy of the License at   *
+ *                                                              *
+ *   http://www.apache.org/licenses/LICENSE-2.0                 *
+ *                                                              *
+ * Unless required by applicable law or agreed to in writing,   *
+ * software distributed under the License is distributed on an  *
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY       *
+ * KIND, either express or implied.  See the License for the    *
+ * specific language governing permissions and limitations      *
+ * under the License.                                           *
+ ****************************************************************/
+
+package org.apache.james.imapserver.netty;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import java.time.Duration;
+
+import org.apache.commons.net.imap.AuthenticatingIMAPClient;
+import org.apache.james.util.concurrency.ConcurrentTestRunner;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+
+@SuppressWarnings("checkstyle:membername")
+class IMAPServerPlainAuthenticateThenAnotherCommandTest extends AbstractIMAPServerTest {
+    IMAPServer imapServer;
+    int port;
+
+    @BeforeEach
+    void setup() throws Exception {
+        imapServer = createImapServer("imapServer.xml");
+        port = imapServer.getListenAddresses().get(0).getPort();
+    }
+
+    @AfterEach
+    void tearDown() {
+        if (imapServer != null) {
+            imapServer.destroy();
+        }
+    }
+
+    @Test
+    void authenticateShouldOnlyConsumeAuthDataCommandNotTheNextCommand() throws Exception {
+        ConcurrentTestRunner.builder()
+                .operation((threadNumber, step) -> {
+                    AuthenticatingIMAPClient imapClient = new AuthenticatingIMAPClient();
+                    imapClient.connect("127.0.0.1", port);
+                    assertThat(imapClient.authenticate(AuthenticatingIMAPClient.AUTH_METHOD.PLAIN, USER.asString(),
+                            USER_PASS)).isTrue();
+                    assertThat(imapClient.logout()).isTrue();
+                    imapClient.disconnect();
+                })
+                .threadCount(10)
+                .operationCount(200)
+                .runSuccessfullyWithin(Duration.ofMinutes(10));
+    }
+}
